backend/requirements.txt
# Web Framework
Flask==2.3.3
Flask-CORS==4.0.0

# AI/ML Libraries
openai-whisper==20231117
torch==2.1.0
transformers==4.35.2
huggingface-hub==0.19.4
sentence-transformers==2.2.2

# Vector Database and RAG
chromadb==0.4.18
langchain==0.0.350
langchain-community>=0.0.2,<0.1.0

# Audio Processing
librosa==0.10.1
soundfile==0.13.1
pydub==0.25.1

# Text-to-Speech (Alternative approach - skip problematic TTS packages initially)
# TTS==0.22.0  # Commented out due to pandas conflict and C++ build issues
# coqui-tts>=0.22.1  # Commented out due to pandas conflict and C++ build issues

# Natural Language Processing
scikit-learn==1.3.2
nltk==3.8.1
spacy==3.7.2

# Email
email-validator==2.1.0

# Utilities
numpy==1.25.2
pandas==1.5.3  # Changed from 2.1.4 to avoid TTS compatibility issues
requests==2.31.0
python-dotenv==1.0.0
pydantic==2.5.0

# File Processing
python-multipart==0.0.6

# Logging and Monitoring
colorlog==6.8.0

# Development
pytest==7.4.3
black==23.11.0
flake8==6.1.0

# Audio format support
ffmpeg-python==0.2.0

# JSON and data handling
orjson==3.9.10

# Additional dependencies that might be needed
numba>=0.56.0
tiktoken>=0.3.0
tokenizers>=0.14.0,<0.19.0
torchvision
sentencepiece
more-itertools

# Alternative TTS solution (lightweight)
gtts==2.4.0  # Google Text-to-Speech as alternative
pygame==2.5.2  # For audio playback